+++ services/settings/pkg/store/defaults/defaults.go
@@ -501,9 +501,6 @@ func DefaultRoleAssignments(cfg *config.Config) []*settingsmsg.UserRoleAssignmen
 	return []*settingsmsg.UserRoleAssignment{
 		// default admin users
 		{
-			AccountUuid: "058bff95-6708-4fe5-91e4-9ea3d377588b", // demo user "moss"
-			RoleId:      BundleUUIDRoleAdmin,
-		}, {
 			AccountUuid: cfg.AdminUserID,
 			RoleId:      BundleUUIDRoleAdmin,
 		},
@@ -520,6 +517,9 @@ func DefaultRoleAssignments(cfg *config.Config) []*settingsmsg.UserRoleAssignmen
 		},
 		// default users with role "spaceadmin"
 		{
+			AccountUuid: "058bff95-6708-4fe5-91e4-9ea3d377588b", // demo user "moss"
+			RoleId:      BundleUUIDRoleAdmin,
+		}, {
 			AccountUuid: "534bb038-6f9d-4093-946f-133be61fa4e7",
 			RoleId:      BundleUUIDRoleSpaceAdmin,
 		},